refactor: refactor code structure
Organize the code and separation of concerns
