+++ internal/groq/config.go
@@ -56,17 +56,3 @@ func NewSystem(message, name string) *System {
 		Name:    name,
 	}
 }
-
-const (
-	DEEPSEEK        = "deepseek-r1-distill-llama-70b"
-	LLAMA           = "llama3-70b-8192"
-	LLAMA_VERSATILE = "llama-3.3-70b-versatile"
-	GEMMA           = "gemma2-9b-it"
-)
-
-var AllowedModels = []string{
-	DEEPSEEK,
-	LLAMA,
-	LLAMA_VERSATILE,
-	GEMMA,
-}
