--- internal/groq/serializer.go
@@ -1,1 +1,0 @@
-package groq
